data/lgs_soru_engine_v3/lgs_soru_engine_v1/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
orjson==3.10.7
python-dotenv==1.0.1
rapidfuzz==3.9.7
numpy==2.0.2
scikit-learn==1.5.2
faiss-cpu==1.8.0.post1
pyyaml==6.0.1
